requirements.txt
numpy~=1.26.4
python-dotenv~=1.2.1
langchain~=0.3.7
langchain-openai~=0.3.0
faiss-cpu~=1.9.0
langchain_community~=0.3.7
bs4~=0.0.2
beautifulsoup4~=4.14.2
openai~=1.59.0
python-iso639~=2024.10.22
pydantic~=2.12.5
tenacity~=8.2.3
chromadb~=1.3.5
pandas~=2.2.3
urllib3~=2.5.0
protobuf~=4.25.3